use proper django forms, added puzzle creation, added meta info

+++ puzzles/views.py
@@ -1,48 +1,42 @@
 from django.shortcuts import get_object_or_404, render
-from django.http import HttpResponse, HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, Http404
 from django.urls import reverse
 
 from .models import Hunt, Round, Puzzle, PuzzleStatus, Answer
+from .forms import NewPuzzleForm, UpdatePuzzleForm
 
 def bigboard(request, hunt_id):
     hunt = get_object_or_404(Hunt, pk=hunt_id)
 
-    return render(request, "puzzles/bigboard.html", {"hunt": hunt})
+    return render(request, "puzzles/bigboard.html", {
+        "hunt": hunt,
+        "new_puzzle_form": NewPuzzleForm()
+    })
 
 def puzzlepage(request, puzzle_id):
     puzzle = get_object_or_404(Puzzle, pk=puzzle_id)
-
     return render(request, "puzzles/puzzlepage.html", {
         "puzzle": puzzle,
-        "puzzle_statuses": PuzzleStatus
+        "update_puzzle_form": UpdatePuzzleForm(initial={
+            "name": puzzle.name,
+            "status": puzzle.status,
+            "url": puzzle.url,
+            "is_meta": puzzle.is_meta
+        })
     })
 
 def update(request, puzzle_id):
     puzzle = get_object_or_404(Puzzle, pk=puzzle_id)
     
-    # TODO: proper validation
+    if request.method == "POST":
+        form = UpdatePuzzleForm(request.POST)
+        if form.is_valid():
+            form.update_puzzle(puzzle)
 
-    # Update puzzle status
-    new_status = request.POST["status"]
-    puzzle.status = new_status
-
-    new_url = request.POST["puzzle_url"]
-    puzzle.url = new_url
-
-    puzzle.save()
-
-    # Delete any requested answer
-    for answer in puzzle.answer_set.all():
-        if request.POST.get(f"delete_answer_{answer.id}", False):
-            answer.delete()
-
-    # Add new answer
-    new_answer_text = request.POST["new_answer"]
-    if new_answer_text != "" and len(new_answer_text) < 255:
-        ans = Answer()
-        ans.answer_text = new_answer_text.upper()
-        ans.puzzle = puzzle
-        ans.save() 
+        # todo: there is definitely a better way to handle answer deletion 
+        for answer in puzzle.answer_set.all():
+            if request.POST.get(f"delete_answer_{answer.id}", False):
+                answer.delete()
 
     return HttpResponseRedirect(reverse("puzzles:puzzlepage", args=(puzzle.id,)))
 
@@ -57,4 +51,21 @@ def new_round(request, hunt_id):
         rnd.hunt = hunt
         rnd.save()
 
+    return HttpResponseRedirect(reverse("puzzles:bigboard", args=(hunt.id,)))
+
+
+def new_puzzle(request, hunt_id, round_id):
+    hunt = get_object_or_404(Hunt, pk=hunt_id)
+    hunt_round = get_object_or_404(Round, pk=round_id)
+
+    if hunt_round.hunt.id != hunt.id:
+        raise Http404("This round is for a different hunt than the given hunt")
+
+    if request.method == "POST":
+        form = NewPuzzleForm(request.POST)
+        if form.is_valid():
+            puzzle = form.make_puzzle()
+            puzzle.hunt_round = hunt_round
+            puzzle.save()
+    
     return HttpResponseRedirect(reverse("puzzles:bigboard", args=(hunt.id,)))
